# Performance improvements for WMA, FWMA and Linreg